requirements.txt
Flask==3.0.3
qrcode[pil]==7.4.2
cryptography==42.0.8
Pillow==10.4.0
schedule==1.2.2
python-dotenv==1.0.1
openpyxl==3.1.5
gunicorn==22.0.0
